you should be able to specify a dir for templates and have all of those templates filled out rather than be limited to a single file

+++ extract.py
@@ -60,7 +60,7 @@ background_color_param_name = "--background"
 background_color_param_default = "auto"
 
 template_param_name = "--template"
-template_param_default = "./templates/iterm.itermcolors"
+template_param_default = "./templates"
 
 config = {
     background_color_param_name: background_color_param_default,
@@ -202,39 +202,51 @@ for img_file_path in image_paths:
         colors[name + "-blue-float"]  = rgb[2] / 255
         colors[name + "-hex"]         = hex
 
-    template_file_path = os.path.realpath(config[template_param_name])
-    # template_file_name = template_file_path.split('/')[-1]
-    # template_file_name_parts = template_file_name.split('.')
-    # template_file_extension = ""
-    # if len(template_file_name_parts) > 1:
-    #     template_file_extension = "." + "".join(template_file_name_parts[1:])
-
     print("=========== Terminal Colors ===========")
     with open('templates/columns-with-headers.txt', 'r') as print_template:
         print(pystache.render(print_template.read(), colors))
     print("=======================================")
 
-    with open(template_file_path, 'r') as template_file:
-        template = template_file.read()
-        template_file_name = os.path.basename(template_file_path)
+    image_name = os.path.basename(img_file_path).split(".")[0]
+    image_extension = os.path.basename(img_file_path).split(".")[-1]
 
-        # set name constant for the iTerm dynamic profile
-        image_name = os.path.basename(img_file_path).split(".")[0]
-        image_extension = os.path.basename(img_file_path).split(".")[-1]
+    output_dir_path = os.path.join(os.path.expanduser("~/.gvcci/themes"), image_name)
+    try:
+        os.stat(output_dir_path)
+    except:
+        os.mkdir(output_dir_path)        
 
-        output_dir_path = os.path.join(os.path.expanduser("~/.gvcci/themes"), image_name)
-        output_theme_path = os.path.join(output_dir_path, template_file_name)
-        try:
-            os.stat(output_dir_path)
-        except:
-            os.mkdir(output_dir_path)
+    output_image_path = os.path.join(output_dir_path, "wallpaper." + image_extension)
+    copyfile(img_file_path, output_image_path)
+    
+    print("Output: " + output_image_path)
 
-        output_image_path = os.path.join(output_dir_path, "wallpaper." + image_extension)
-        copyfile(img_file_path, output_image_path)
+    template_file_or_dir_path = os.path.realpath(config[template_param_name])
+    if (os.path.isfile(template_file_or_dir_path)):
+        template_file_path = template_file_or_dir_path
+        with open(template_file_path, 'r') as template_file:
+            template = template_file.read()
+            template_file_name = os.path.basename(template_file_path)
+            
+            output_theme_path = os.path.join(output_dir_path, template_file_name)
 
-        with open(output_theme_path, 'w') as out_file:
-            out_file.write(pystache.render(template, colors))
-        
-        print("Output: " + output_theme_path)
-        print("Output: " + output_image_path)
+
+            with open(output_theme_path, 'w') as out_file:
+                out_file.write(pystache.render(template, colors))
+            
+            print("Output: " + output_theme_path)
+    elif (os.path.isdir(template_file_or_dir_path)):
+        for template_file in os.listdir(template_file_or_dir_path):
+            template_file_path = os.path.join(template_file_or_dir_path, template_file)
+            with open(template_file_path, 'r') as template_file:
+                template = template_file.read()
+                template_file_name = os.path.basename(template_file_path)
+                
+                output_theme_path = os.path.join(output_dir_path, template_file_name)
+
+
+                with open(output_theme_path, 'w') as out_file:
+                    out_file.write(pystache.render(template, colors))
+                
+                print("Output: " + output_theme_path)
